// frontend/src/lib/api/generated/api.ts
export interface UnitResponseDto {
    'id'?: string;
    'tenantId'?: string;
    'propertyId'?: string;
    'structureId'?: string;
    'unitNumber'?: string;
    'status'?: UnitResponseDtoStatusEnum;
    'specification'?: string;
    'baselinePrice'?: number;
    'currencyCode'?: string;
    'createdAt'?: string;
    'updatedAt'?: string;
}

export const UnitResponseDtoStatusEnum = {
    Vacant: 'VACANT',
    Occupied: 'OCCUPIED',
    Maintenance: 'MAINTENANCE',
} as const;

export type UnitResponseDtoStatusEnum = typeof UnitResponseDtoStatusEnum[keyof typeof UnitResponseDtoStatusEnum];